requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
bcrypt==3.2.0
black==21.9b0
cffi==1.14.6
click==8.0.1
constantly==15.1.0
cryptography==35.0.0
DateTime==4.3
dnspython==2.1.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.68.1
fastapi-users==8.1.0
flake8==3.9.2
h11==0.12.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Jinja2==3.0.1
makefun==1.11.3
MarkupSafe==2.0.1
mccabe==0.6.1
motor==2.5.1
mypy-extensions==0.4.3
openpyxl==3.0.9
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.3
regex==2021.9.30
six==1.16.0
srv==1.1.0
srvlookup==2.0.0
starlette==0.14.2
tomli==1.2.1
Twisted==21.7.0
typing-extensions==3.10.0.2
ujson==4.2.0
uvicorn==0.15.0
zope.interface==5.4.0
